CC-1909: Refactor to consolidate all default values for api.Options into api/options.go
